# packfile: improve Index memory representation to be more compact